add tests, update readme,add utils directory
